Moved ClientMessageEncoder next to the ClientMessageDecoder
These 2 classes should be placed next to each other since they
are 2 sides of the same coin.
